refactor: isolate mobx helpers from shared guards and expand renderer coverage tests
