+++ src/tests/setup/registerRenderersForTests.ts
@@ -1,0 +1,4 @@
+// Some specs exercise registrations without loading `IgniteElementFactory`
+// directly, so we import the built-in strategies here to populate the registry.
+import "../../renderers/ignite-jsx";
+import "../../renderers/lit";
