+++ src/utils/adapterGuards.ts
@@ -1,5 +1,4 @@
 import type { EnhancedStore, Slice } from "@reduxjs/toolkit";
-import { isObservable } from "mobx";
 import type { AnyStateMachine } from "xstate";
 
 export interface XStateActorLike {
@@ -78,31 +77,6 @@ export function isReduxSlice(source: unknown): source is Slice {
 		candidate.actions !== null &&
 		typeof candidate.getInitialState === "function"
 	);
-}
-
-export function isMobxObservable(source: unknown): boolean {
-	if (typeof source !== "object" || source === null) {
-		return false;
-	}
-
-	if (isObservable(source)) {
-		return true;
-	}
-
-	// biome-ignore lint/suspicious/noPrototypeBuiltins: supporting older JS targets without Object.hasOwn
-	if (Object.prototype.hasOwnProperty.call(source, "$$observable")) {
-		return true;
-	}
-	// biome-ignore lint/suspicious/noPrototypeBuiltins: supporting older JS targets without Object.hasOwn
-	if (Object.prototype.hasOwnProperty.call(source, "_atom")) {
-		return true;
-	}
-	// biome-ignore lint/suspicious/noPrototypeBuiltins: supporting older JS targets without Object.hasOwn
-	if (Object.prototype.hasOwnProperty.call(source, "$mobx")) {
-		return true;
-	}
-
-	return false;
 }
 
 export function isFunction<T extends (...args: unknown[]) => unknown>(
